refactor(runtime): move alias to runtime html, remove old bb vc exports in runtime
