+++ packages/__tests__/1-kernel/module-loader.spec.ts
@@ -6,8 +6,7 @@ import * as ce_a_b from './modules/ce-a-b.js';
 import * as ce_default from './modules/ce-default.js';
 import * as kitchen_sink from './modules/kitchen-sink.js';
 import { assert } from '@aurelia/testing';
-import { CustomAttribute, CustomElement } from '@aurelia/runtime-html';
-import { BindingBehavior, ValueConverter } from '@aurelia/runtime';
+import { CustomAttribute, CustomElement, BindingBehavior, ValueConverter } from '@aurelia/runtime-html';
 
 describe('ModuleLoader', function () {
   function createFixture() {
