+++ packages/__tests__/3-runtime-html/binding-commands.spec.ts
@@ -1,5 +1,5 @@
-import { alias } from '@aurelia/runtime';
 import {
+  alias,
   CommandType,
   BindingCommandInstance,
   bindingCommand,
