+++ packages/__tests__/3-runtime-html/custom-attributes.spec.ts
@@ -1,5 +1,5 @@
-import { alias } from '@aurelia/runtime';
 import {
+  alias,
   bindable,
   customAttribute,
   INode,
