+++ packages/__tests__/3-runtime-html/repeater-custom-element.spec.ts
@@ -1,6 +1,6 @@
 import { Class } from '@aurelia/kernel';
-import { BindingBehavior } from '@aurelia/runtime';
 import {
+  BindingBehavior,
   customElement,
   bindable,
   CustomElement,
