+++ packages/__tests__/3-runtime-html/value-converters.spec.ts
@@ -1,8 +1,6 @@
 import {
   alias,
-  valueConverter
-} from '@aurelia/runtime';
-import {
+  valueConverter,
   bindable,
   customAttribute,
   INode,
