+++ packages/runtime-html/src/resources/custom-attribute.ts
@@ -1,12 +1,11 @@
 import { mergeArrays, firstDefined, Key } from '@aurelia/kernel';
-import { BindingMode, registerAliases } from '@aurelia/runtime';
 import { Bindable } from '../bindable';
 import { Watch } from '../watch';
 import { getRef } from '../dom';
 import { DefinitionType } from './resources-shared';
 import { appendResourceKey, defineMetadata, getAnnotationKeyFor, getOwnMetadata, getResourceKeyFor, hasOwnMetadata } from '../utilities-metadata';
 import { isFunction, isString } from '../utilities';
-import { aliasRegistration, transientRegistration } from '../utilities-di';
+import { aliasRegistration, registerAliases, transientRegistration } from '../utilities-di';
 
 import type {
   Constructable,
@@ -19,6 +18,7 @@ import type {
 import type { BindableDefinition, PartialBindableDefinition } from '../bindable';
 import type { ICustomAttributeViewModel, ICustomAttributeController } from '../templating/controller';
 import type { IWatchDefinition } from '../watch';
+import { BindingMode } from '@aurelia/runtime';
 
 declare module '@aurelia/kernel' {
   interface IContainer {
